Added function authenticate_user
this function checks if a user exists in the users file and then
checks to see if the saved password is the same as the password
entered by the user on login. If the user exists and the password is
correct the the function returns a Fernet object with a derived_key to
be used in the login session.