Implementing persistent login
Have created a function that creates a session file that stores the
authenticated user's session information temporarily and securely.
The goals are:
Avoid prompting for the master password repeatedly during short-term use.
Persist the authenticated user and encryption key securely between runs.
Automatically use the session if it exists, and is valid.

It is now possible to log in and a session file is created. The list
and get_password functions are not working correctly yet however.